requirements.txt
aiogram==3.6.0
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.4
python-dateutil==2.9.0.post0
Pillow==10.4.0
pytesseract>=0.3.10
openai>=1.30.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
python-dotenv==1.0.0
telethon==1.34.0
aiohttp==3.9.1
opencv-python>=4.8.0
numpy>=1.24.0
APScheduler==3.10.4